# doc/openapi.json
{
  "info": {
    "version": "0.1.0", 
...
    "title": "EThing HTTP API"
  }, 
  "paths": {
...
    "/api/resources": {
      "get": {
        "tags": [
          "resource"
        ], 
        "responses": {
          "200": {
            "description": "A list of resources", 
            "schema": {
              "items": {
                "$ref": "#/definitions/Resource"
              }, 
              "type": "array"
            }
          }
        }, 
        "parameters": [
          {
            "x-nullable": true, 
            "name": "q", 
            "in": "query", 
            "default": null, 
            "required": false, 
            "type": "string", 
            "description": "Query string for searching resources"
          }, 
          {
            "minimum": 0, 
            "name": "skip", 
            "in": "query", 
            "format": "int32", 
            "required": false, 
            "type": "integer", 
            "description": "Skips a number of resources"
          }, 
          {
            "minimum": 0, 
            "name": "limit", 
            "in": "query", 
            "format": "int32", 
            "required": false, 
            "type": "integer", 
            "description": "Limits the number of resources returned"
          }, 
          {
            "name": "fields", 
            "in": "query", 
            "items": {
              "type": "string"
            }, 
            "required": false, 
            "type": "string", 
            "collectionFormat": "multi"
          }, 
          {
            "description": "The key on which to do the sorting, by default the sort is made by modifiedDate descending. To make the sort descending, prepend the field name by minus \"-\". For instance, \"-createdDate\" will sort by createdDate descending", 
            "required": false, 
            "type": "string", 
            "name": "sort", 
            "in": "query"
          }
        ], 
        "description": "Lists the resources.\n\n#### cURL example\n\n```bash\ncurl -H 'X-API-KEY: <YOUR_API_KEY>' http://localhost:8000/api/resources\n```"
      }
    }, 
...
    "/api/devices": {
      "post": {
        "tags": [
          "device"
        ], 
        "responses": {
          "200": {
            "description": "The device was successfully created", 
            "schema": {
              "allOf": [
                {
                  "required": [
                    "type"
                  ], 
                  "type": "object", 
                  "properties": {
                    "type": {
                      "type": "string", 
                      "description": "The type of the device to create (eg: \"Http\" or \"MySensorsEthernetGateway\")."
                    }
                  }
                }, 
                {
                  "$ref": "#/definitions/Device"
                }
              ]
            }
          }
        }, 
        "parameters": [
          {
            "required": true, 
            "in": "body", 
            "description": "\nThe metadata of the device to be created.\n\nexample:\n\n```json\n{\n   \"name\": \"mydevice.txt\",\n   \"location\":{\n      \"latitude\": 5.12,\n      \"longitude\": -45.78\n   },\n   \"scope\": \"resource:read notification\",\n}\n```\n", 
            "name": "metadata", 
            "schema": {
              "$ref": "#/definitions/Device"
            }
          }, 
          {
            "name": "fields", 
            "in": "query", 
            "items": {
              "type": "string"
            }, 
            "required": false, 
            "type": "string", 
            "collectionFormat": "multi"
          }
        ], 
        "description": "Creates a new device."
      }
    }, 
...
    "/api/devices/{r}/call/{operationId}": {
      "post": {
        "tags": [
          "device"
        ], 
        "responses": {
          "200": {
            "description": "The response of the device."
          }
        }, 
        "parameters": [
          {
            "required": true, 
            "type": "string", 
            "description": "id of the resource. Devices or Apps using the api key authentication may use the word 'me' to replace their id.", 
            "in": "path", 
            "name": "id"
          }, 
          {
            "description": "id of the operation.", 
            "required": true, 
            "in": "path", 
            "type": "string", 
            "name": "operationId"
          }, 
          {
            "required": false, 
            "in": "body", 
            "description": "required parameters for this operation.", 
            "name": "data", 
            "schema": {
              "type": "object"
            }
          }, 
          {
            "description": "An id representing a Resource", 
            "required": true, 
            "type": "string", 
            "name": "r", 
            "in": "path"
          }
        ], 
        "description": "Execute an operation identified by operationId. The parameters can either be passed in the query string or in the body as a JSON object or a x-www-form-urlencoded string."
      }, 
      "get": {
        "tags": [
          "device"
        ], 
        "responses": {
          "200": {
            "description": "The response of the device."
          }
        }, 
        "parameters": [
          {
            "description": "id of the operation.", 
            "required": true, 
            "in": "path", 
            "type": "string", 
            "name": "operationId"
          }, 
          {
            "description": "An id representing a Resource", 
            "required": true, 
            "type": "string", 
            "name": "r", 
            "in": "path"
          }
        ], 
        "description": "Execute an operation identified by operationId. The parameters must be passed in the query string."
      }
    }, 
    "/api/devices/{r}/api/{operationId}": {
      "get": {
        "tags": [
          "device"
        ], 
        "responses": {
          "200": {
            "description": "object describing the operation."
          }
        }, 
        "parameters": [
          {
            "description": "id of the operation.", 
            "required": true, 
            "in": "path", 
            "type": "string", 
            "name": "operationId"
          }, 
          {
            "description": "An id representing a Resource", 
            "required": true, 
            "type": "string", 
            "name": "r", 
            "in": "path"
          }
        ], 
        "description": "Retrieves an object describing the operation identified by operationId."
      }
    }, 
    "/api/devices/{r}/api": {
      "get": {
        "tags": [
          "device"
        ], 
        "responses": {
          "200": {
            "description": "object describing the operations available for this device."
          }
        }, 
        "parameters": [
          {
            "description": "An id representing a Resource", 
            "required": true, 
            "type": "string", 
            "name": "r", 
            "in": "path"
          }
        ], 
        "description": "Retrieves an object describing the operations available for this device."
      }
    }, 
...
    "/api/resources/{r}": {
      "get": {
        "tags": [
          "resource"
        ], 
        "responses": {
          "200": {
            "description": "resource object", 
            "schema": {
              "$ref": "#/definitions/Resource"
            }
          }
        }, 
        "parameters": [
          {
            "description": "An id representing a Resource", 
            "required": true, 
            "type": "string", 
            "name": "r", 
            "in": "path"
          }, 
          {
            "name": "fields", 
            "in": "query", 
            "items": {
              "type": "string"
            }, 
            "required": false, 
            "type": "string", 
            "collectionFormat": "multi"
          }
        ], 
        "description": "Returns the meta-data of a resource in JSON."
      }, 
      "delete": {
        "tags": [
          "resource"
        ], 
        "responses": {
          "200": {
            "description": "the resource has been deleted successfully"
          }
        }, 
        "parameters": [
          {
            "description": "An id representing a Resource", 
            "required": true, 
            "type": "string", 
            "name": "r", 
            "in": "path"
          }, 
          {
            "default": false, 
            "required": false, 
            "type": "boolean", 
            "name": "children", 
            "in": "query"
          }
        ], 
        "description": "deletes a resource"
      }, 
      "patch": {
        "tags": [
          "resource"
        ], 
        "responses": {
          "200": {
            "description": "resource successfully updated", 
            "schema": {
              "$ref": "#/definitions/Resource"
            }
          }
        }, 
        "parameters": [
          {
            "required": true, 
            "in": "body", 
            "description": "the attributes to modify", 
            "name": "modification", 
            "schema": {
              "$ref": "#/definitions/Resource"
            }
          }, 
          {
            "description": "An id representing a Resource", 
            "required": true, 
            "type": "string", 
            "name": "r", 
            "in": "path"
          }, 
          {
            "name": "fields", 
            "in": "query", 
            "items": {
              "type": "string"
            }, 
            "required": false, 
            "type": "string", 
            "collectionFormat": "multi"
          }
        ], 
        "description": "update a resource. Only properties which are not readonly can be modified.\n\nRename a resource :\n\n```json\n{\n   \"name\":\"myFileRenamed.txt\"\n}\n```\n\nClear a description :\n\n```json\n{\n   \"description\":null\n}\n```"
      }
    }, 
...
  }, 
  "schemes": [
    "http"
  ], 
...
  "produces": [
    "application/json"
  ], 
  "securityDefinitions": {
    "api_key_query": {
      "in": "query", 
      "type": "apiKey", 
      "description": "authentication through an API key, used only by devices or apps.", 
      "name": "api_key"
    }, 
    "api_key": {
      "in": "header", 
      "type": "apiKey", 
      "description": "authentication through an API key, used only by devices or apps.", 
      "name": "X-API-KEY"
    }, 
    "basic_auth": {
      "type": "basic", 
      "description": "basic authentication."
    }
  }, 
  "definitions": {
...
    "Device": {
      "allOf": [
        {
          "$ref": "#/definitions/Resource"
        }, 
        {
          "type": "object", 
          "properties": {
            "methods": {
              "items": {}, 
              "readOnly": true, 
              "type": "array", 
              "description": "The list of the methods available."
            }, 
            "battery": {
              "default": null, 
              "type": "null", 
              "description": "The battery level of this device (must be between 0 (empty) and 100 (full) , or null if the device has no battery information)."
            }, 
            "interfaces": {
              "items": {}, 
              "readOnly": true, 
              "type": "array", 
              "description": "A list of intefaces this device inherit"
            }, 
            "connected": {
              "default": false, 
              "type": "boolean", 
              "description": "Set to true when this device is connected."
            }, 
            "location": {
              "default": null, 
              "type": "null", 
              "description": "The location of this device."
            }, 
            "lastSeenDate": {
              "default": null, 
              "type": "null", 
              "description": "Last time this device was reached or made a request."
            }
          }
        }
      ]
    }, 
...
    "Resource": {
      "discriminator": "type", 
      "required": [
        "name"
      ], 
      "type": "object", 
      "properties": {
        "description": {
          "default": "", 
          "type": "string", 
          "description": "A description of this resource."
        }, 
        "data": {
          "additionalProperties": {
            "type": "string"
          }, 
          "default": {}, 
          "type": "object", 
          "properties": {}, 
          "description": "A collection of arbitrary key-value pairs. Entries with null values are cleared in update. The keys must not be empty or longer than 64 characters, and must contain only the following characters : letters, digits, underscore and dash. Values must be either a string or a boolean or a number"
        }, 
        "public": {
          "default": false, 
          "enum": [
            false, 
            "readonly", 
            "readwrite"
          ], 
          "description": "False: this resource is not publicly accessible. 'readonly': this resource is accessible for reading by anyone. 'readwrite': this resource is accessible for reading and writing by anyone."
        }, 
        "extends": {
          "items": {}, 
          "readOnly": true, 
          "type": "array", 
          "description": "An array of classes this resource is based on."
        }, 
        "modifiedDate": {
          "default": "2018-04-03T15:04:13.953939", 
          "type": "string", 
          "description": "Last time this resource was modified", 
          "format": "date-time"
        }, 
        "createdBy": {
          "default": null, 
          "type": "null", 
          "description": "The id of the resource responsible of the creation of this resource, or null."
        }, 
        "createdDate": {
          "readOnly": true, 
          "type": "string", 
          "description": "Create time for this resource", 
          "format": "date-time"
        }, 
        "type": {
          "readOnly": true, 
          "type": "string", 
          "description": "The type of the resource"
        }, 
        "id": {
          "readOnly": true, 
          "type": "string", 
          "description": "The id of the resource"
        }, 
        "name": {
          "minLength": 1, 
          "type": "string", 
          "description": "The name of the resource", 
          "pattern": "^[a-zA-Z0-9 !#$%&'()+,\\-.;=@^_`{    ]+(\\/[a-zA-Z0-9 !#$%&'()+,\\-.;=@^_`{    ]+)*$"
        }
      }, 
      "description": "The base representation of a resource object"
    }, 
...
  }, 
  "security": [
    {
      "api_key": []
    }, 
    {
      "api_key_query": []
    }, 
    {
      "basic_auth": []
    }
  ], 
  "swagger": "2.0", 
  "consumes": [
    "application/json"
  ]
}